imwrite 관련 오류 수정

+++ ImageProcessor.cpp
@@ -48,7 +48,7 @@ torch::Tensor ImageProcessor::post_process(cv::Mat image){
     current_time_t = std::chrono::system_clock::to_time_t(current_time);
     std::string img_path;
     img_path.append("examples/").append(std::ctime(&current_time_t)).append(".jpg");
-    //imwrite(img_path, image);
+    imwrite(img_path, image);
 
     float MEAN[] = {0.48897059, 0.46548275, 0.4294};
     float STD[] = {0.22861765, 0.22948039, 0.24054667};
@@ -92,14 +92,17 @@ void ImageProcessor::process_image(cv::Mat& image){//시간 나면 std::vector<c
     output = module.forward({input_tensor}).toTensor();
     image = tensor_to_image(output);
 
-
-    //imwrite 버그있어서 고쳐야함.
-    /*
     if (!img_path.empty())
         img_path.clear();
     img_path.append("examples/").append(std::ctime(&current_time_t)).append(".png");
-    imwrite(img_path, image);
-    */
+    if (std::filesystem::exists(img_path)) {
+        size_t lastDotIndex = img_path.find_last_of(".");
+        std::string filename = img_path.substr(0, lastDotIndex);
+        filename.append("_").append("png");
+        imwrite(filename, image);
+    } else {
+        imwrite(img_path, image);
+    }
 }
 
 //img_path의 이미지를 로드함
